% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cars.R
\docType{data}
\name{cars}
\alias{cars}
\title{a dataset of horsepower for different cars}
\format{
A data frame.
}
\usage{
data(cars)
}
\description{
First 4 columns of the mtcars dataset.

\itemize{
\item mpg. miles per gallon
\item cyl. cylinders
\item disp. displacement
\item hp. horse poser
}
}
\references{
R base package.
}
\keyword{datasets}
