% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-iris2.R
\docType{data}
\name{iris2}
\alias{iris2}
\title{sepal length for differen species of iris}
\format{
A data frame.
}
\usage{
data(iris2)
}
\description{
The Species and Sepal.Length from the iris data set from base R.

\itemize{
\item Species. species
\item Sepal.Length. sepal length
}
}
\references{
from R base package.
}
\keyword{datasets}
